Made the operation_Name default to the name of the request (#85)
In C#, application insights will automatically set the operation_Name to e.g. GET /

With appinsights-rs, the operation_Name is empty by default. An empty operation_name does not show up in the application insights dashboard. I think it is most often desired to set operation_Name to the value of name that is set already.